Fixed modifiers selection
[Why]
With how this was patched previously, we changed into the single player version of GameplaySetupView which is bad for mods that have different views for multi and also it carried over the saved modifiers for single player. The goal is to have the multiplayer modifiers view show up as a tab in GameplaySetup while keeping everything intact.

[How]
Harmony and zenject to the rescue! A static event is called when we enter in the level selection, subscribed to from our zenjected MultiplayerGSViewController which replaces the "Multiplayer" tab with the "Modifiers" tab. This modifiers tab has the multiplayer modifiers view which we get through zenject and when we exit, another event is triggered through harmony and this zenject object hides the multiplayer modifiers and we go back to the regular multiplayer tab. Now we could achieve the enhanced UI but also keeping the implementation near stock. Also, clicking on the modifiers tab does nothing since it is not interactable, it is only to show modifiers and players could just click into the level button. Did this because it can create confusion when going into levels by clicking into modifiers button.